# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Wishaw YMCA Badge API
  version: 0.0.1
  description: |
    OpenAPI contract for the Wishaw YMCA Esports badging backend.

    This API is designed for a separate React frontend and uses session-based authentication.
    Successful login establishes a JSESSIONID cookie; subsequent calls must include that cookie.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/admin/progress/legacy-points:
    post:
      tags:
        - Progress
      summary: Set imported or legacy badge points for a player
      operationId: setLegacyPoints
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLegacyPointsRequest'
      responses:
        '200':
          description: Updated badge progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadgeProgressResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/admin/progress/award-challenge:
    post:
      tags:
        - Progress
      summary: Award a challenge to a player and recalculate progress
      operationId: awardChallenge
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AwardChallengeRequest'
      responses:
        '200':
          description: Updated badge progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadgeProgressResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session cookie returned by POST /api/v1/auth/login.
...
  responses:
...
    ValidationError:
      description: Bean validation failed for one or more request fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    ServerError:
      description: An unexpected backend error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    SetLegacyPointsRequest:
      type: object
      required:
        - playerId
        - badgeCategoryId
        - legacyPoints
      properties:
        playerId:
          type: integer
          format: int64
        badgeCategoryId:
          type: integer
          format: int64
        legacyPoints:
          type: integer
          minimum: 0
    AwardChallengeRequest:
      type: object
      required:
        - playerId
        - challengeId
      properties:
        playerId:
          type: integer
          format: int64
        challengeId:
          type: integer
          format: int64
        notes:
          type: string
          nullable: true
...
    BadgeProgressResponse:
      type: object
      properties:
        badgeCategoryId:
          type: integer
          format: int64
        categoryCode:
          type: string
        categoryName:
          type: string
        legacyPoints:
          type: integer
        earnedPoints:
          type: integer
        totalPoints:
          type: integer
        currentLevel:
          type: string
          nullable: true
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...